Refactoring to maintain project standards
